Updated basic protocol, added a simple System handler, which needs to be implemented on the server so we can have basic client connection, and then basic client and channel components need to be implemented before any useful features are provided.

+++ jchat_client/include/core/chat_component.h
@@ -27,7 +27,7 @@ public:
 
   // Handler functions
   virtual ComponentType GetType() = 0;
-  virtual bool Handle(ChatClient &client, uint8_t message_type,
+  virtual bool Handle(ChatClient &client, uint16_t message_type,
     TypedBuffer &buffer) = 0;
 };
 }
